+++ modules/client/src/multisig/internal/constants.ts
@@ -15,6 +15,14 @@ export const AVAILABLE_FUNCTION_SIGNATURES: string[] = [
   ).format("minimal"),
 ];
 
+// TODO:
+// use abi from plugin metadata
+export const UPDATE_ABI: { [build: number]: MetadataAbiInput[] } = {
+  1: [],
+};
+
+// TODO:
+// use abi from plugin metadata
 export const INSTALLATION_ABI: MetadataAbiInput[] = [
   {
     internalType: "address[]",