+++ modules/client/test/integration/addresslistVoting-client/methods.test.ts
@@ -27,6 +27,7 @@ import {
 import {
   ADDRESS_ONE,
   contextParamsLocalChain,
+  IPFS_CID,
   SUBGRAPH_ACTIONS,
   SUBGRAPH_PROPOSAL_BASE,
   SUBGRAPH_VOTERS,
@@ -57,10 +58,13 @@ import {
 import {
   Context,
   PrepareInstallationStep,
+  PrepareUpdateStep,
   ProposalMetadata,
   ProposalStatus,
   SortDirection,
 } from "@aragon/sdk-client-common";
+import { PluginRepo__factory } from "@aragon/osx-ethers";
+import { createAddresslistVotingPluginBuild } from "../../helpers/create-plugin-build";
 
 describe("Client Address List", () => {
   let server: Server;
@@ -364,6 +368,106 @@ describe("Client Address List", () => {
             expect(typeof step.versionTag.build).toBe("number");
             expect(typeof step.versionTag.release).toBe("number");
             break;
+        }
+      }
+    });
+  });
+
+  describe("Plugin update", () => {
+    it("Should update prepare a plugin update for a tokenVotingPlugin", async () => {
+      const ctx = new Context(contextParamsLocalChain);
+      const client = new AddresslistVotingClient(ctx);
+
+      const { dao, plugin } = await buildAddressListVotingDAO(
+        repoAddr,
+      );
+      const release = 1;
+      await createAddresslistVotingPluginBuild(
+        release,
+        deployment.addresslistVotingRepo.address,
+      );
+
+      const pluginSetupInstance = PluginRepo__factory.connect(
+        deployment.addresslistVotingRepo.address,
+        client.web3.getConnectedSigner(),
+      );
+      const version = await pluginSetupInstance["getLatestVersion(uint8)"](
+        release,
+      );
+
+      expect(version.tag.release).toBe(release);
+      expect(version.tag.build).toBe(2);
+
+      const mockedClient = mockedGraphqlRequest.getMockedInstance(
+        client.graphql.getClient(),
+      );
+      const installation = {
+        appliedPreparation: {
+          helpers: [],
+          pluginRepo: {
+            id: deployment.addresslistVotingRepo.address,
+          },
+        },
+        appliedVersion: {
+          metadata: `ipfs://${IPFS_CID}`,
+          release: {
+            release: 1,
+          },
+          build: 1,
+        },
+      };
+      mockedClient.request.mockResolvedValueOnce({
+        iplugin: { installations: [installation] },
+      });
+      const steps = client.methods.prepareUpdate(
+        {
+          daoAddressOrEns: dao,
+          pluginAddress: plugin,
+          newVersion: {
+            release: 1,
+            build: 2,
+          },
+        },
+      );
+      for await (const step of steps) {
+        switch (step.key) {
+          case PrepareUpdateStep.PREPARING:
+            expect(typeof step.txHash).toBe("string");
+            expect(step.txHash).toMatch(/^0x[A-Fa-f0-9]{64}$/i);
+            break;
+          case PrepareUpdateStep.DONE:
+            expect(typeof step.pluginAddress).toBe("string");
+            expect(step.pluginAddress).toBe(plugin);
+            expect(step.initData instanceof Uint8Array).toBe(true);
+            expect(step.pluginAddress).toMatch(/^0x[A-Fa-f0-9]{40}$/i);
+            expect(typeof step.pluginRepo).toBe("string");
+            expect(step.pluginRepo).toBe(
+              deployment.addresslistVotingRepo.address,
+            );
+            expect(step.pluginRepo).toMatch(/^0x[A-Fa-f0-9]{40}$/i);
+            expect(typeof step.versionTag.build).toBe("number");
+            expect(step.versionTag.build).toBe(2);
+            expect(typeof step.versionTag.release).toBe("number");
+            expect(step.versionTag.release).toBe(1);
+            for (const permission of step.permissions) {
+              if (permission.condition) {
+                expect(typeof permission.condition).toBe("string");
+                expect(permission.condition).toMatch(
+                  /^0x[A-Fa-f0-9]{40}$/i,
+                );
+              }
+              expect(typeof permission.operation).toBe("number");
+              expect(typeof permission.where).toBe("string");
+              expect(permission.where).toMatch(/^0x[A-Fa-f0-9]{40}$/i);
+              expect(typeof permission.who).toBe("string");
+              expect(permission.who).toMatch(/^0x[A-Fa-f0-9]{40}$/i);
+            }
+            break;
+          default:
+            throw new Error(
+              "Unexpected DAO prepare update step: " +
+                JSON.stringify(step, null, 2),
+            );
         }
       }
     });