+++ modules/client/test/integration/client/encoding.test.ts
@@ -33,6 +33,7 @@ import { keccak256 } from "@ethersproject/keccak256";
 import { AddressZero } from "@ethersproject/constants";
 import {
   ApplyInstallationParams,
+  ApplyUpdateParams,
   Context,
   DaoAction,
   TokenType,
@@ -571,6 +572,78 @@ describe("Client", () => {
         );
       }
     });
+    it("Should encode an applyUpdate action", async () => {
+      const context = new Context(contextParamsLocalChain);
+      const client = new Client(context);
+
+      const applyUpdateParams: ApplyUpdateParams = {
+        permissions: [{
+          operation: 1,
+          permissionId: PermissionIds.EXECUTE_PERMISSION_ID,
+          where: "0x1234567890123456789012345678901234567890",
+          who: "0x2345678901234567890123456789012345678901",
+        }],
+        versionTag: {
+          build: 1,
+          release: 1,
+        },
+        pluginRepo: "0x2345678901234567890123456789012345678901",
+        pluginAddress: "0x1234567890123456789012345678901234567890",
+        initData: new Uint8Array([0, 1, 2, 3]),
+        helpers: [],
+      };
+      const daoAddress = "0x1234567890123456789012345678901234567890";
+      const actions = client.encoding.applyUpdateAction(
+        daoAddress,
+        applyUpdateParams,
+      );
+
+      expect(actions.length).toBe(3);
+      expect(typeof actions[1]).toBe("object");
+      expect(actions[1].data).toBeInstanceOf(Uint8Array);
+
+      const daoInterface = PluginSetupProcessor__factory.createInterface();
+      const hexString = bytesToHex(actions[1].data);
+      const argsDecoded = daoInterface.decodeFunctionData(
+        "applyUpdate",
+        hexString,
+      );
+      expect(argsDecoded.length).toBe(2);
+      expect(argsDecoded[0]).toBe(
+        daoAddress,
+      );
+      expect(argsDecoded[1].pluginSetupRef.versionTag.build).toBe(
+        applyUpdateParams.versionTag.build,
+      );
+      expect(argsDecoded[1].pluginSetupRef.versionTag.release).toBe(
+        applyUpdateParams.versionTag.release,
+      );
+      expect(argsDecoded[1].plugin).toBe(
+        applyUpdateParams.pluginAddress,
+      );
+      expect(argsDecoded[1].pluginSetupRef.pluginSetupRepo).toBe(
+        applyUpdateParams.pluginRepo,
+      );
+      for (const index in argsDecoded[1].permissions) {
+        expect(argsDecoded[1].permissions[parseInt(index)].operation).toBe(
+          applyUpdateParams.permissions[parseInt(index)].operation,
+        );
+        expect(argsDecoded[1].permissions[parseInt(index)].where).toBe(
+          applyUpdateParams.permissions[parseInt(index)].where,
+        );
+        expect(argsDecoded[1].permissions[parseInt(index)].who).toBe(
+          applyUpdateParams.permissions[parseInt(index)].who,
+        );
+        expect(argsDecoded[1].permissions[parseInt(index)].condition).toBe(
+          AddressZero,
+        );
+        expect(argsDecoded[1].permissions[parseInt(index)].permissionId).toBe(
+          applyUpdateParams.permissions[parseInt(index)]
+            .permissionId,
+        );
+      }
+    });
+
     it("Should encode an initializeFrom action", async () => {
       const context = new Context(contextParamsLocalChain);
       const client = new Client(context);