+++ modules/client/test/integration/client/estimation.test.ts
@@ -9,7 +9,12 @@ import {
   DepositParams,
   SetAllowanceParams,
 } from "../../../src";
-import { contextParamsLocalChain } from "../constants";
+import {
+  ADDRESS_ONE,
+  ADDRESS_THREE,
+  ADDRESS_TWO,
+  contextParamsLocalChain,
+} from "../constants";
 import * as ganacheSetup from "../../helpers/ganache-setup";
 import * as deployContracts from "../../helpers/deployContracts";
 import { Server } from "ganache";
@@ -28,6 +33,8 @@ describe("Client", () => {
       server = await ganacheSetup.start();
       deployment = await deployContracts.deploy();
       contextParamsLocalChain.daoFactoryAddress = deployment.daoFactory.address;
+      contextParamsLocalChain.pluginSetupProcessorAddress =
+        deployment.pluginSetupProcessor.address;
       contextParamsLocalChain.ensRegistryAddress =
         deployment.ensRegistry.address;
     });
@@ -179,6 +186,26 @@ describe("Client", () => {
       expect(typeof gasFeesEstimation.max).toEqual("bigint");
       expect(gasFeesEstimation.max).toBeGreaterThan(BigInt(0));
       expect(gasFeesEstimation.max).toBeGreaterThan(gasFeesEstimation.average);
+    });
+    it("Should estimate the gas fees for preparing an update", async () => {
+      const ctx = new Context(contextParamsLocalChain);
+      const client = new Client(ctx);
+
+      const estimation = await client.estimation.prepareUpdate({
+        pluginAddress: ADDRESS_ONE,
+        pluginRepo: ADDRESS_TWO,
+        daoAddressOrEns: ADDRESS_THREE,
+        newVersion: {
+          build: 1,
+          release: 2,
+        },
+      });
+
+      expect(typeof estimation).toEqual("object");
+      expect(typeof estimation.average).toEqual("bigint");
+      expect(typeof estimation.max).toEqual("bigint");
+      expect(estimation.max).toBeGreaterThan(BigInt(0));
+      expect(estimation.max).toBeGreaterThan(estimation.average);
     });
   });
 });